requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic>=2.7.0
langchain==0.1.13
langchain-core>=0.1.40
langchain-community==0.0.29
langchain-groq==0.1.0
chromadb==0.4.24
beautifulsoup4==4.12.3
requests==2.31.0
python-multipart==0.0.9
PyJWT>=2.10.1
passlib==1.7.4
bcrypt==4.1.2
httpx==0.27.0
websockets>=13.0
chromadb==0.4.24
sentence-transformers==2.5.1

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
supabase==2.28.0
PyPDF2==3.0.0
pandas==2.2.1
langchain-text-splitters==0.0.1
pydub==0.25.1
audioop-lts==0.2.2
fpdf2==2.8.1
resend==2.1.0
